refactor(utils): text_helper와 vocab_dict 분리
